poll-api/requirements.txt
kafka-python==2.0.2
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==1.3.1